// src/features/reframing/enums.ts
export enum ReframingModalityLabel {
  COGNITIVE_REFRAMING = "Cognitive Reframing",
  STOIC_REFRAMING = "Stoic Reframing",
  EXISTENTIAL_REFRAMING = "Existential Reframing",
  SPIRITUAL_REFRAMING = "Spiritual Reframing",
  NARRATIVE_REFRAMING = "Narrative Reframing",
  POSITIVE_PSYCHOLOGY = "Positive Psychology",
}

export enum ReframingModalityIdentifier {
  COGNITIVE_REFRAMING = "cognitive_reframing",
  STOIC_REFRAMING = "stoic_reframing",
  EXISTENTIAL_REFRAMING = "existential_reframing",
  SPIRITUAL_REFRAMING = "spiritual_reframing",
  NARRATIVE_REFRAMING = "narrative_reframing",
  POSITIVE_PSYCHOLOGY = "positive_psychology",
}